+++ .gitignore
@@ -1,2 +1,1 @@
-dist/
 node_modules/
