+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "esbuild src/main.ts --bundle --platform=node --outfile=./dist/main.js --watch --servedir=. --serve=3000 --loader:.frag=text --loader:.vert=text",
+    "dev": "esbuild src/main.ts --bundle --platform=node --outfile=./dist/main.js --watch --servedir=dist --serve=3000 --loader:.frag=text --loader:.vert=text",
     "build": "esbuild src/main.ts --bundle --outfile=./dist/main.js --minify --loader:.frag=text --loader:.vert=text --platform=node"
   },
   "repository": {
