+++ src/render_manager.ts
